+++ R/3_1_baseline_data.R
@@ -1,4 +1,7 @@
-## BASELINE TAB ##
+### Baseline Data Tab ###
+
+
+# Static UI ---------------------------------------------------------------
 
 baselineDataUI <- function() {
   ns <- NS("baseline")
@@ -8,6 +11,10 @@ baselineDataUI <- function() {
     uiOutput(ns("mainUI")) %>% withSpinnerProxy(),
   )
 }
+
+
+
+# Server ------------------------------------------------------------------
 
 #' requires global data frame 'baseline_data'
 #' @param cur_stn a `reactive()` single line data frame
@@ -149,6 +156,18 @@ baselineDataServer <- function(cur_stn, has_focus) {
       ) %>% rowwise()
 
       ## stnSummaryData ----
+      makeMinMax <- function(df, var) {
+        v <- df[[var]]
+        if (length(v) == 0) return(tibble())
+        tibble(
+          observations = length(na.omit(v)),
+          min = df[which.min(v), ][[var]],
+          max = df[which.max(v), ][[var]],
+          date_of_min = df[which.min(v), ]$date,
+          date_of_max = df[which.max(v), ]$date
+        )
+      }
+
       stnSummaryData <- reactive({
         req(input$year)
         req(selected_data_ready())
@@ -157,7 +176,7 @@ baselineDataServer <- function(cur_stn, has_focus) {
           distinct(date, .keep_all = T)
         date_fmt <- ifelse(input$year == "All", "%b %e, %Y", "%b %e")
         baseline_summary_vars %>%
-          reframe(pick(everything()), make_min_max(df, var)) %>%
+          reframe(pick(everything()), makeMinMax(df, var)) %>%
           mutate(across(c(min, max), ~paste(.x, units))) %>%
           mutate(across(c(date_of_min, date_of_max), ~format(.x, date_fmt))) %>%
           select(-c(var, units)) %>%
@@ -172,16 +191,19 @@ baselineDataServer <- function(cur_stn, has_focus) {
       output$viewDataUI <- renderUI({
         req(input$year)
 
-        btn_year <- downloadButton(ns("downloadYear"), paste("Download", input$year, "data"))
-        btn_all <- downloadButton(ns("downloadAll"), "Download all years of baseline data for this site")
-        dl_btns <- { if (input$year == "All") list(btn_all) else list(btn_year, btn_all) }
-
         tagList(
-          p(dl_btns),
-          div(
-            style = "overflow: auto;",
-            dataTableOutput(ns("dataTable"))
-          )
+          p(
+            strong("Station ID:"), cur_stn()$station_id, br(),
+            strong("Station Name:"), cur_stn()$station_name, br(),
+            strong("Waterbody:"), cur_stn()$waterbody
+          ),
+          p(
+            if (input$year != "All")
+              downloadButton(ns("downloadStnYear"), sprintf("Download station data (%s)", input$year)),
+            downloadButton(ns("downloadStn"), "Download station data (all years)"),
+            downloadButton(ns("downloadBaseline"), "Download entire baseline dataset")
+          ),
+          div(style = "overflow: auto;", dataTableOutput(ns("dataTable")))
         )
       })
 
@@ -190,7 +212,6 @@ baselineDataServer <- function(cur_stn, has_focus) {
         req(selected_data_ready())
 
         date_fmt <- ifelse(input$year == "All", "%b %d, %Y", "%b %d")
-
         df <- selected_data() %>%
           arrange(date) %>%
           distinct(date, .keep_all = T) %>%
@@ -199,24 +220,31 @@ baselineDataServer <- function(cur_stn, has_focus) {
           mutate(across(everything(), as.character)) %>%
           pivot_longer(cols = -label, names_to = "Parameter") %>%
           pivot_wider(names_from = label) %>%
-          mutate(Parameter = gsub("D o", "D.O.", Parameter)) %>%
+          mutate(Parameter = gsub("D o", "DO", Parameter)) %>%
           mutate(Parameter = gsub("P h", "pH", Parameter))
 
         datatable(df, selection = "none", options = list(paging = F))
       },
         server = F)
 
-      ## downloadYear ----
-      output$downloadYear <- downloadHandler(
-        paste0("stn-", cur_stn()$station_id, "-baseline-data-", input$year, ".csv"),
-        function(file) { write_csv(selected_data(), file) }
+      ## downloadStnYear ----
+      output$downloadStnYear <- downloadHandler(
+        sprintf("WAV Stn %s Baseline Data (%s).csv", cur_stn()$station_id, input$year),
+        function(file) { write_csv(selected_data(), file, na = "") }
       )
 
-      ## downloadAll ----
-      output$downloadAll <- downloadHandler(
-        paste0("stn-", cur_stn()$station_id, "-baseline-data.csv"),
-        function(file) { write_csv(cur_data(), file) }
+      ## downloadStn ----
+      output$downloadStn <- downloadHandler(
+        sprintf("WAV Stn %s Baseline Data.csv", cur_stn()$station_id),
+        function(file) { write_csv(cur_data(), file, na = "") }
       )
+
+      ## downloadBaseline ----
+      output$downloadBaseline <- downloadHandler(
+        "WAV Baseline Data.csv",
+        function(file) { write_csv(baseline_data, file, na = "") }
+      )
+
 
       # Return values ----
       return(reactive(list(year = input$year)))